[feat]Camera Component 추가 / ItemBox InventoryUI 추가
- Camera Component 추가
- ItemBox InventoryUI 추가
