+++ Assets/DEV/YJE/BoxTestScene.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 96c4db6c8fb806b4898b54db50f40f98
-PrefabImporter:
+guid: 72bad928b12b2ed4b8a55340c99aeb83
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
