+++ Assets/DEV/YJE/CameraController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b9b1d35a645dd67468915a3951e17e43
+guid: 20cb3b31e4f04dc46a79468f7e415741
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
