+++ Assets/DEV/YJE/ItmeBoxInventory.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cacd8a9a778e2974298781599c73f1fa
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
