+++ Assets/DEV/YJE/PlayerCameraController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b235a5210be8cf245bc7ac4cda74c822
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
