+++ Assets/DEV/YJE/Prefabs/ClearShot Camera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7c25b100e3ad6e649a07b8b79a8ac50f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
