+++ Assets/DEV/YJE/Scripts/BoxController.cs
@@ -6,12 +6,14 @@ using UnityEngine;
 public class BoxController : MonoBehaviour
 {
     public bool IsUIOpen;
+    [SerializeField] GameObject UI_ItemBox;
 
     // TODO : UI 상호작용 창 닫혀있는지 확인하는 bool변수 - public
     public void BoxOpen()
     {
         Debug.Log("아이템상자 열기");
         IsUIOpen = true;
+        UI_ItemBox.SetActive(true);
         // TODO : UI 상호작용 창 닫혀있는지 확인하는 bool변수 = ture; - return값
     }
 
@@ -19,6 +21,8 @@ public class BoxController : MonoBehaviour
     {
         Debug.Log("아이템상자 닫기");
         IsUIOpen = false;
+        UI_ItemBox.SetActive(false);
+
         // TODO : UI 상호작용 창 닫혀있는지 확인하는 bool변수 = false; - return값
 
     }
