+++ Assets/DEV/YJE/Scripts/ResourceController.cs.meta
