+++ Assets/DEV/YJE/Test.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 844aaabf31aa2354a81dc68bcbd100f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
