+++ Assets/DEV/YJE/Test/TestCam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f53bd1129683a2459de8c60cc48668a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
