+++ Assets/DEV/YJE/Test/TestClearShot Camera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 032b7fccac3fb0541afcb94ef7edb14a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
