+++ Assets/DEV/YJE/Test/TestGameScene.cs
@@ -71,6 +71,7 @@ public class TestGameScene : MonoBehaviourPunCallbacks
     {
         //플레이어를 생성할 랜덤한 위치
         Vector3 randomPos = new Vector3(Random.Range(-3f, 3f), 0, Random.Range(-3f, 3f));
-        nowPlayer = PhotonNetwork.Instantiate("JHS/Player01", randomPos, Quaternion.identity);
+        //nowPlayer = PhotonNetwork.Instantiate("JHS/Player01", randomPos, Quaternion.identity);
+        nowPlayer = PhotonNetwork.Instantiate("YJE/TestPlayer", randomPos, Quaternion.identity);
     }
 }
