+++ Assets/Resources/JHS/Player01.prefab
@@ -10182,6 +10182,7 @@ MonoBehaviour:
   hunger: 0
   warmthMax: 500
   warmth: 0
+  animator: {fileID: 0}
   playerDie: 0
   ishungry: 0
   iscold: 0
